+++ packages/rise/rise/rise.py
@@ -9,10 +9,9 @@ from pygeoapi.provider.base import BaseProvider
 from pygeoapi.util import crs_transform
 from com.env import TRACER
 from rise.lib.cache import RISECache
-from com.geojson.types import GeojsonFeatureCollectionDict
+from com.geojson.types import GeojsonFeatureCollectionDict, SortDict
 from rise.lib.location import LocationResponseWithIncluded
 from rise.lib.types.location import LocationDataAttributes
-from rise.lib.types.sorting import SortDict
 
 LOGGER = logging.getLogger(__name__)
