+++ packages/snotel/snotel/lib/locations.py
@@ -3,10 +3,9 @@
 
 from com.cache import RedisCache
 from com.env import TRACER
-from com.geojson.types import GeojsonFeature, GeojsonFeatureCollectionDict
+from com.geojson.types import GeojsonFeature, GeojsonFeatureCollectionDict, SortDict
 from com.helpers import await_, parse_bbox, parse_z
 import geojson_pydantic
-import geojson_pydantic.types
 from rise.lib.types.helpers import ZType
 from snotel.lib.types import StationDTO
 import shapely
@@ -53,12 +52,15 @@ class LocationCollection:
         return self._filter_by_geometry(geometry, z)
 
     def to_geojson(
-        self, skip_geometry: Optional[bool] = False, itemsIDSingleFeature=False
+        self,
+        skip_geometry: Optional[bool] = False,
+        itemsIDSingleFeature=False,
+        sortby: Optional[list[SortDict]] = None,
     ) -> GeojsonFeatureCollectionDict | GeojsonFeature:
         """
         Return a geojson feature if the client queried for items/{itemId} or a feature collection if they queried for items/ even if the result is only one item
         """
-        features: list[GeojsonFeature] = []
+        features: list[geojson_pydantic.Feature] = []
 
         for loc in self.locations:
             feature: GeojsonFeature = {
@@ -72,21 +74,23 @@ class LocationCollection:
                 else None,
                 "id": loc.stationId,
             }
-            features.append(feature)
+            features.append(geojson_pydantic.Feature.model_validate(feature))
 
         geojson_pydantic.FeatureCollection(
             type="FeatureCollection",
-            features=[geojson_pydantic.Feature.model_validate(f) for f in features],
+            features=features,
         )
         if itemsIDSingleFeature:
             assert len(features) == 1, (
                 "The user queried a single item but we have more than one present. This is a sign that filtering by locationid wasn't done properly"
             )
-            return features[0]
-        return {
-            "type": "FeatureCollection",
-            "features": features,
-        }
+            return GeojsonFeature(**features[0].model_dump())
+        return GeojsonFeatureCollectionDict(
+            **{
+                "type": "FeatureCollection",
+                "features": features,
+            }
+        )
 
     @TRACER.start_as_current_span("geometry_filter")
     def _filter_by_geometry(
